pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "regmixer"
dynamic = ["version"]
readme = "README.md"
description = ""
authors = [
    { name = "Allen Institute for Artificial Intelligence", email = "oe-data-engineering@allenai.org" }
]
requires-python = ">=3.9"
license = { file = "LICENSE" }
dependencies = [
    "ai2-olmo-core @ git+https://github.com/allenai/OLMo-core.git@1a218ba90001bb5f9f082c183dcf13a0e2f732ce",
    "boto3",
    "click",
    "pydantic",
    "s3fs",
    "yaspin",
    "gcsfs",
    "pyarrow<21",
    "datasets>=3,<4",
    "tokenizers>=0.22,<1"
]
[project.optional-dependencies]
dev = [
    "ruff",
    "mypy>=1.0,<1.4",
    "black>=23.1,<24.0",
    "isort>=5.12,<5.13",
    "pytest",
]
beaker = [
    "beaker-py>=1,<2",
    "GitPython>=3.0,<4.0",
]
wandb = [
    "wandb",
]
eval = [
    "lightgbm",
    "pandas",
    "scipy",
    "seaborn",
    "statsmodels"
]
all = [
    "regmixer[dev,beaker,wandb,eval]",
]

[project.scripts]
rmc = "regmixer.cli:cli"
rmc-eval = "regmixer.eval.cli:cli"
rmc-workspace = "regmixer.workspace:cli"
rmc-internal = "regmixer.internal.cli:cli"
rmc-convert = "regmixer.internal.convert_mixture:cli"
rmc-convert-cookbook = "regmixer.internal.convert_cookbook_mixture:cli"

[tool.black]
line-length = 100
include = '\.pyi?$'
exclude = '''
(
      __pycache__
    | \.git
    | \.mypy_cache
    | \.pytest_cache
    | \.vscode
    | \.venv
    | \bdist\b
    | \bdoc\b
    | scratch/
    | build/
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.ruff]
line-length = 115

[tool.ruff.lint]
ignore = ["F403", "F405", "E501"]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".venv",
    "venv",
    ".mypy_cache",
    "__pycache__",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "doc",
    "pretrain_data",
    "inference",
]

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F401"]

[tool.pyright]
reportPrivateImportUsage = false

[tool.mypy]
ignore_missing_imports = true
no_site_packages = true
check_untyped_defs = true
disable_error_code = "has-type"

[[tool.mypy.overrides]]
module = "tests.*"
strict_optional = false

[tool.pytest.ini_options]
testpaths = "tests/"
python_classes = [
  "Test*",
  "*Test",
]
log_format = "%(asctime)s - %(levelname)s - %(name)s - %(message)s"
log_level = "DEBUG"
log_cli = false
log_cli_level = "DEBUG"
filterwarnings = [
    'ignore::FutureWarning:huggingface_hub\.file_download',
    'ignore::DeprecationWarning:pkg_resources',
    'ignore::DeprecationWarning:google\.rpc',
    'ignore::FutureWarning:torch\.distributed\.checkpoint\.default_planner',
]


